Improve build-cross; add build-source
Improve build-cross: icons and desktop files are now included, as well
as a more Nix-like bin/ and share/ structure. This should make it easier
to package for Nix and other binary distros.

Add build-source: there's now an action to build a source tarball that
also contains the vendor/ directory. This should make it easier to
package for Nix and other source distros. This fixes issue #1.

Usage:

    cat<<EOF | nix-build -E "$(cat)"
    import ./nix {
        action = "build-source";
        rev = "$(git describe --tags)";
    }
    EOF